backend/requirements.txt
torch
torchvision==0.20.0
fastapi[standard]
pillow